Split up lineage command into subcommand files
